pyproject.toml
[project]
name = "spcal"
version = "0.0.9"
description = "A toolchain to calculate sp of a cell"
authors = [{ name = "Yaqi Sun" }]
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "pyyaml>=6.0.2,<7",
    "typer>=0.13.0,<0.14",
    "networkx>=3.4.2,<4",
    "matplotlib>=3.9.2,<4",
    "plotly>=5.24.1,<6",
    "yaspin>=3.1.0,<4",
    "schemdraw>=0.19,<0.20",
    "scipy>=1.15.1,<2",
    "regex>=2024.11.6,<2025",
    "pydantic>=2.11.3,<3",
    "graphviz>=0.20.3,<0.21",
    "dd>=0.6.0,<0.7",
    "pytest>=8.4.1",
    "black>=25.9.0",
    "autoflake>=2.3.1",
    "isort>=7.0.0",
    "py-spy>=0.4.1",
    "nuitka>=2.8.4",
]

[tool.hatch.build.targets.sdist]
include = ["src/spcal"]

[tool.hatch.build.targets.wheel]
include = ["src/spcal"]

[tool.hatch.build.targets.wheel.sources]
"src/spcal" = "spcal"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
